Implemented under-sampling for train DataFrame for class balance
This commit introduces a refined version of the `under_sampling` function, which selectively removes 3/4th of the groups with a 'label' value of 0 for each unique 'sub' in the DataFrame. This targeted under-sampling approach is designed to address class imbalance issues, particularly reducing the over-representation of a specific class (label 0) in the dataset. The function maintains the integrity of the data for other labels, ensuring a more balanced dataset for model training and evaluation.